build: move constant from indext to constants to avoid circular import
