+++ packages/salesforcedx-vscode-core/src/decorators/scratch-org-decorator.ts
@@ -7,7 +7,7 @@
 
 import { ConfigUtil } from '@salesforce/salesforcedx-utils-vscode';
 import { StatusBarAlignment, StatusBarItem, window } from 'vscode';
-import { ORG_OPEN_COMMAND } from '../../src';
+import { ORG_OPEN_COMMAND } from '../../src/constants';
 import { nls } from '../messages';
 
 let statusBarItem: StatusBarItem | undefined;
