+++ packages/salesforcedx-vscode-core/test/jest/decorators/scratch-org-decorator.test.ts
@@ -1,6 +1,6 @@
 import { ConfigUtil } from '@salesforce/salesforcedx-utils-vscode';
 import * as vscode from 'vscode';
-import { ORG_OPEN_COMMAND } from '../../../src';
+import { ORG_OPEN_COMMAND } from '../../../src/constants';
 import { showOrg } from '../../../src/decorators';
 import { nls } from '../../../src/messages';
 
@@ -24,13 +24,14 @@ describe('scratch org decorator', () => {
       onDidChange: jest.fn(),
       onDidCreate: jest.fn()
     };
-    createStatusBarItemMock = (vscode.window.createStatusBarItem as any).mockReturnValue(
-      mockStatusBarItem
-    );
+    createStatusBarItemMock = (vscode.window
+      .createStatusBarItem as any).mockReturnValue(mockStatusBarItem);
     createFileSystemWatcherMock = (vscode.workspace
       .createFileSystemWatcher as any).mockReturnValue(mockWatcher);
-    getDefaultUsernameOrAliasMock = jest
-      .spyOn(ConfigUtil, 'getDefaultUsernameOrAlias');
+    getDefaultUsernameOrAliasMock = jest.spyOn(
+      ConfigUtil,
+      'getDefaultUsernameOrAlias'
+    );
   });
 
   describe('show Org', () => {
@@ -39,7 +40,9 @@ describe('scratch org decorator', () => {
       await showOrg();
       expect(getDefaultUsernameOrAliasMock).toHaveBeenCalled();
       expect(createStatusBarItemMock).toHaveBeenCalled();
-      expect(mockStatusBarItem.tooltip).toEqual(nls.localize('status_bar_open_org_tooltip'));
+      expect(mockStatusBarItem.tooltip).toEqual(
+        nls.localize('status_bar_open_org_tooltip')
+      );
       expect(mockStatusBarItem.command).toEqual(openOrgCommand);
       expect(mockStatusBarItem.show).toHaveBeenCalled();
       expect(mockStatusBarItem.text).toEqual(browserIcon);
